refactor: create service
Convert `resolveDependencies` function into service that needs to be injected.
